creates CoroutinesTest
