+++ purchases/src/test/java/com/revenuecat/purchases/CoroutinesTest.kt
@@ -1,0 +1,62 @@
+package com.revenuecat.purchases
+
+import androidx.test.ext.junit.runners.AndroidJUnit4
+import io.mockk.verify
+import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.test.runTest
+import org.assertj.core.api.Assertions
+import org.junit.Test
+import org.junit.runner.RunWith
+import org.robolectric.annotation.Config
+
+@OptIn(ExperimentalCoroutinesApi::class)
+@RunWith(AndroidJUnit4::class)
+@Config(manifest = Config.NONE)
+@Suppress("DEPRECATION")
+class CoroutinesTest : BasePurchasesTest() {
+
+    @Test
+    fun `retrieve customer info - Success`() = runTest {
+        mockCustomerInfoHelper()
+
+        val result = purchases.getCustomerInfoSuspend()
+
+        verify(exactly = 1) {
+            mockCustomerInfoHelper.retrieveCustomerInfo(
+                appUserId,
+                any(),
+                any(),
+                any(),
+            )
+        }
+        Assertions.assertThat(result is Result.Success).isTrue
+    }
+
+    @Test
+    fun `retrieve customer info - Error`() = runTest {
+        mockCustomerInfoHelper(PurchasesError(PurchasesErrorCode.CustomerInfoError, "Customer info error"))
+
+        val result = purchases.getCustomerInfoSuspend()
+
+        verify(exactly = 1) {
+            mockCustomerInfoHelper.retrieveCustomerInfo(
+                appUserId,
+                any(),
+                any(),
+                any(),
+            )
+        }
+        Assertions.assertThat(result is Result.Error).isTrue
+    }
+
+    @Test
+    fun `retrieve customer info - CustomerInfoError`() = runTest {
+        mockCustomerInfoHelper(PurchasesError(PurchasesErrorCode.CustomerInfoError, "Customer info error"))
+
+        val result = purchases.getCustomerInfoSuspend() as Result.Error
+
+        Assertions.assertThat(result.value.code).isEqualTo(PurchasesErrorCode.CustomerInfoError)
+    }
+
+    // endregion
+}
